feat(transport): add mineral_type and camelCase-alias ShipmentOut
Adds Shipment.mineral_type (the field the frontend mock has but the
model didn't) and applies the same camelCase-alias convention Sites
already uses, including lotId/mineral aliases.

Also fixes a real bug found while testing this against a live token:
Pydantic v2 serializes Decimal fields as JSON strings by default, not
numbers, which silently broke every Decimal-typed field's contract
with the frontend's `number` types (including Phase 1's SiteOut,
fixed here too — lat/lng/gradePct/confidence). Switched every *Out
schema's numeric fields from Decimal to float; input schemas keep
Decimal for write-side precision.

+++ app/sites/schemas.py
@@ -11,12 +11,12 @@ class SiteOut(BaseModel):
     id: str
     name: str
     district: str
-    lat: Decimal
-    lng: Decimal
+    lat: float
+    lng: float
     primaryMineral: str = Field(validation_alias="primary_mineral")
     secondaryMinerals: list = Field(validation_alias="secondary_minerals")
-    gradePct: Decimal = Field(validation_alias="grade_pct")
-    confidence: Decimal
+    gradePct: float = Field(validation_alias="grade_pct")
+    confidence: float
     estimatedTonnage: int = Field(validation_alias="estimated_tonnage")
     safetyScore: int = Field(validation_alias="safety_score")
     riskLevel: str = Field(validation_alias="risk_level")
@@ -31,8 +31,8 @@ class SiteCreate(BaseModel):
     name: str
     district: str = ""
     country_code: str = "RW"
-    lat: Decimal
-    lng: Decimal
+    lat: float
+    lng: float
     primary_mineral: str
     secondary_minerals: list = []
     grade_pct: Decimal = Decimal("0")
